+++ packages/swarm-styles/CHANGELOG.md
@@ -3,6 +3,39 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.6.0](https://github.com/meetup/swarm-ui/compare/@meetup/swarm-styles@0.5.19...@meetup/swarm-styles@0.6.0) (2019-07-17)
+
+
+### Bug Fixes
+
+* resolved focus issue, ForwardedButton for forwardRefs in dropdown ([f4829ba](https://github.com/meetup/swarm-ui/commit/f4829ba))
+* updated and added MenuLink to demo ([b0f780a](https://github.com/meetup/swarm-ui/commit/b0f780a))
+
+
+### Features
+
+* dropdown wip, buggy initial focus ([716a5cb](https://github.com/meetup/swarm-ui/commit/716a5cb))
+
+
+
+
+
+## [0.5.19](https://github.com/meetup/swarm-ui/compare/@meetup/swarm-styles@0.5.18...@meetup/swarm-styles@0.5.19) (2019-07-16)
+
+**Note:** Version bump only for package @meetup/swarm-styles
+
+
+
+
+
+## [0.5.18](https://github.com/meetup/swarm-ui/compare/@meetup/swarm-styles@0.5.17...@meetup/swarm-styles@0.5.18) (2019-07-09)
+
+**Note:** Version bump only for package @meetup/swarm-styles
+
+
+
+
+
 ## [0.5.17](https://github.com/meetup/swarm-ui/compare/@meetup/swarm-styles@0.5.16...@meetup/swarm-styles@0.5.17) (2019-06-07)
 
 **Note:** Version bump only for package @meetup/swarm-styles